requirements.txt
aerich==0.2.2
aiohttp==3.6.2
anyio==1.4.0
asgiref==3.3.0
async-generator==1.10
async-timeout==3.0.1
asyncclick==7.0.9
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
Django==3.1.3
idna==2.10
iso8601==0.1.12
multidict==4.7.6
#psycopg2==2.8.6
psycopg2-binary
pydantic==1.6.1
pylibmc==1.6.1
pytz==2020.4
requests==2.24.0
sniffio==1.1.0
sqlparse==0.4.1
tomlkit==0.7.0
typing-extensions==3.7.4.2
urllib3==1.25.10
yarl==1.5.1

